Add support for Intel AES-NI.
The implementation is based off Intel's AES whitepaer and Crypto++'s public
domain implementation.  Since this is a first pass implementation, it does not
do hardware assisted key expansion yet.  Additionally with more code, GCM and
CTR can go even faster.

Note that this only works on GCC (and compatible) compilers and it's neccecary
to edit both config.h and the Makefile to use the module.

Visual Studio technically supports the aes intrinsics, so it probably will work
with a bit of poking at the code, but I don't care enough to do it.